handle legacy Next.js versions
