+++ packages/next/package.json
@@ -9,6 +9,7 @@
     "test": "jest --reporters=default --reporters=jest-junit --env node --verbose --bail --runInBand --testTimeout=360000",
     "test-unit": "pnpm test test/unit/",
     "test-next-local": "pnpm test test/integration/*.test.js test/integration/*.test.ts",
+    "test-next-local-legacy": "pnpm test test/integration/integration-legacy-tests.js",
     "test-next-local:middleware": "pnpm test test/integration/middleware.test.ts",
     "test-e2e": "rm -f test/builder-info.json; pnpm test test/fixtures/**/*.test.js",
     "type-check": "tsc --noEmit"
