+++ packages/next/test/integration/integration-1.test.js
@@ -174,21 +174,6 @@ if (parseInt(process.versions.node.split('.')[0], 10) >= 16) {
     expect(buildResult.output['index']).toBeDefined();
     // expect(buildResult.output['index/index']).toBeDefined();
   });
-
-  it('should show error from basePath with legacy monorepo build', async () => {
-    let error;
-
-    try {
-      await runBuildLambda(path.join(__dirname, 'legacy-monorepo-basepath'));
-    } catch (err) {
-      error = err;
-    }
-    console.error(error);
-
-    expect(error.message).toBe(
-      'basePath can not be used with `builds` in vercel.json, use Project Settings to configure your monorepo instead'
-    );
-  });
 }
 
 it('should build using server build', async () => {
@@ -355,32 +340,6 @@ it('should build custom error lambda correctly', async () => {
   expect(notFoundRoute).toBeTruthy();
 });
 
-it('should build initial beforeFiles rewrites', async () => {
-  const {
-    buildResult: { output, routes },
-  } = await runBuildLambda(
-    path.join(__dirname, 'initial-before-files-rewrite')
-  );
-
-  expect(output['index']).toBeDefined();
-  expect(output['another']).toBeDefined();
-  expect(output['dynamic/[slug]']).toBeDefined();
-  expect(output['fallback/[slug]']).toBeDefined();
-  expect(output['api']).toBeDefined();
-  expect(output['api/another']).toBeDefined();
-  expect(output['api/blog/[slug]']).toBeDefined();
-  expect(output['_app']).not.toBeDefined();
-  expect(output['_error']).not.toBeDefined();
-  expect(output['_document']).not.toBeDefined();
-
-  const rewriteRoute = routes.find(route => {
-    return route.dest === '/somewhere';
-  });
-
-  expect(rewriteRoute.check).toBe(true);
-  expect(rewriteRoute.continue).toBeUndefined();
-});
-
 it('Should build the standard example', async () => {
   const {
     buildResult: { output },
@@ -398,27 +357,6 @@ it('Should build the standard example', async () => {
   expect(hasUnderScoreAppStaticFile).toBeTruthy();
   expect(hasUnderScoreErrorStaticFile).toBeTruthy();
   expect(serverlessError).toBeTruthy();
-});
-
-it('Should build the 404-getstaticprops example', async () => {
-  const { buildResult } = await runBuildLambda(
-    path.join(__dirname, '404-getstaticprops')
-  );
-  const { output } = buildResult;
-
-  expect(output['404']).toBeDefined();
-  expect(output['404'].type).toBe('FileFsRef');
-  expect(output['404'].allowQuery).toBe(undefined);
-  expect(output['_next/data/testing-build-id/404.json']).toBeDefined();
-  expect(output['_next/data/testing-build-id/404.json'].type).toBe('FileFsRef');
-  expect(output['_next/data/testing-build-id/404.json'].allowQuery).toBe(
-    undefined
-  );
-  const filePaths = Object.keys(output);
-  const hasUnderScoreErrorStaticFile = filePaths.some(filePath =>
-    filePath.match(/static.*\/pages\/_error-.*\.js$/)
-  );
-  expect(hasUnderScoreErrorStaticFile).toBeTruthy();
 });
 
 it('Should build the 404-getstaticprops-i18n example', async () => {
@@ -500,69 +438,6 @@ it('Should not deploy preview lambdas for static site', async () => {
   expect(output['dynamic'].type).toBe('Prerender');
   expect(output['dynamic'].allowQuery).toEqual([]);
   expect(output['dynamic'].lambda).toBeDefined();
-});
-
-it('Should opt-out of shared lambdas when routes are detected', async () => {
-  const {
-    buildResult: { output },
-  } = await runBuildLambda(
-    path.join(__dirname, '../fixtures/26-mono-repo-404-lambda')
-  );
-  expect(output['packages/webapp/404']).toBeDefined();
-  expect(output['packages/webapp/index']).toBeDefined();
-  expect(output['packages/webapp/__NEXT_PAGE_LAMBDA_0']).not.toBeDefined();
-  const filePaths = Object.keys(output);
-  const hasUnderScoreAppStaticFile = filePaths.some(filePath =>
-    filePath.match(/static.*\/pages\/_app\.js$/)
-  );
-  const hasUnderScoreErrorStaticFile = filePaths.some(filePath =>
-    filePath.match(/static.*\/pages\/_error\.js$/)
-  );
-  expect(hasUnderScoreAppStaticFile).toBeTruthy();
-  expect(hasUnderScoreErrorStaticFile).toBeTruthy();
-});
-
-it('Should build the monorepo example', async () => {
-  const {
-    buildResult: { output },
-  } = await runBuildLambda(path.join(__dirname, 'monorepo'));
-
-  expect(output['www/index']).not.toBeDefined();
-  expect(output['www/__NEXT_PAGE_LAMBDA_0']).toBeDefined();
-  expect(output['www/static/test.txt']).toBeDefined();
-  expect(output['www/data.txt']).toBeDefined();
-  const filePaths = Object.keys(output);
-  const hasUnderScoreAppStaticFile = filePaths.some(filePath =>
-    filePath.match(/static.*\/pages\/_app\.js$/)
-  );
-  const hasUnderScoreErrorStaticFile = filePaths.some(filePath =>
-    filePath.match(/static.*\/pages\/_error\.js$/)
-  );
-  expect(hasUnderScoreAppStaticFile).toBeTruthy();
-  expect(hasUnderScoreErrorStaticFile).toBeTruthy();
-});
-
-it('Should build the legacy standard example', async () => {
-  const {
-    buildResult: { output },
-  } = await runBuildLambda(path.join(__dirname, 'legacy-standard'));
-  expect(output.index).toBeDefined();
-  const filePaths = Object.keys(output);
-  const hasUnderScoreAppStaticFile = filePaths.some(filePath =>
-    filePath.match(/static.*\/pages\/_app\.js$/)
-  );
-  const hasUnderScoreErrorStaticFile = filePaths.some(filePath =>
-    filePath.match(/static.*\/pages\/_error\.js$/)
-  );
-  expect(hasUnderScoreAppStaticFile).toBeTruthy();
-  expect(hasUnderScoreErrorStaticFile).toBeTruthy();
-});
-
-it('Should build the legacy custom dependency test', async () => {
-  const {
-    buildResult: { output },
-  } = await runBuildLambda(path.join(__dirname, 'legacy-custom-dependency'));
-  expect(output.index).toBeDefined();
 });
 
 it('Should throw when package.json or next.config.js is not the "src"', async () => {
